[Channel] Ensure we've a video title before print it
Summary:
Sometimes, a video from Vimeo doesn't have a title.

Before, "PRIVMSG #channel :" were sent to the IRC server, with
an extraneous command sent to the server.

With D166 and a new prefix, we had a "[Vid] " line printed.

We now check the length of the video title.

Issue reported by popolon.

Test Plan: Tested on Freenode #fauve.

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D191