Notifications: drop services filter
Summary:
We filtered known services to avoid a flood from unhandled events.

As it seems, we have all the more frequent events handled, and to
avoid to amend Wearg configuration each time a new event is fired
to our broker, we drop the filter.

Test Plan: Fire a GitHub fork event

Reviewers: dereckson

Differential Revision: http://devcentral.nasqueron.org/D306