Allow .+prenom and .+nom on #wikidata-fr for every user
Summary:
To protect against potential abuse, the flag D is currently required
to run commands. As #wikidata-fr is dedicated for Wikidata work, and
there is no particular issue on this channel, this access control is
more a blocker for legitimate work than useful to stop vandalism.

This change removes this access control for whitelisted channels,
and keeps it in other channels.

Fixes T903.

[ New registry setting ]

  - wikidata.channels: a list of channels where everyone is allowed to
    use Wikidata commands.

Test Plan:
   - `areWikidataCommandsAllowed` and `isWikidataChannel` tested with
     all the four scenarii (D user or not, in Wikidata channel or not).
   - `areWikidataCommandsAllowed` tested not to query registry when the
     user has the D flag, to avoid a MySQL query.

Reviewers: Ash-Crow, dachary, dereckson

Maniphest Tasks: T903

Differential Revision: https://devcentral.nasqueron.org/D605