Move rabbitmq.so to lib directory
Summary:
Old path was /home/surfboard/rabbitmq.so

New path is /srv/viperserv/lib/rabbitmq.so
(with /srv/viperserv the egg working directory).

This is more coherent with the new server doctrine,
where /home is for human accounts, and services are
installed in /srv or /opt.

Ref T1309.

Test Plan: Run Wearg on Ysul

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1309

Differential Revision: https://devcentral.nasqueron.org/D1208