Initialize broker and notifications
Summary:
After 90 seconds, enough to allow the egg to connect to Freenode
and join every channel, load first the broker, then the notifications
code (the notifications code depends of the broker).

This responsibility was previously for the configuration file.

It's more convenient to move this part to the application code,
especially as the bot-specific configuration file is now standardized.

Ref T1309

Test Plan: Restart Wearg, check it connects to broker

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1309

Differential Revision: https://devcentral.nasqueron.org/D1227