Implement resolve_nick and whois commands
Summary:
The whois command is extracted from the Twitter code to tag !pub messages.
This command produces a known nickname string, or an empty string if unknown.

The resolve_nick will produce a known handle or cloak username if known,
or keep the original nickname if unknown. This is intended to be used
with the servers log on quieter channels with mature legitimate users.

Test Plan: Tested on Wearg

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D1322